mnist/requirements.txt
# Core
keras>=2.2
tensorflow-cpu>=2.2.0
scipy<=1.2.0